Added basic map rendering (Textureless)
Basic stuff. Reads in a "map" file (for a temp global 2d array) that it
will then draw to a surface and store in as a variable. This surface
will then be blitted onto the display surface and displayed. I have also
added a basic README file (that doesn't say much)

Also added a changelog file.